resturcture files, add initDatabase command
