+++ election/election.go
@@ -14,22 +14,3 @@ func New(db *database.Database) *Election {
 		db: db,
 	}
 }
-
-// Candidate represents one candidate to vote on
-type Candidate struct {
-	Name   string
-	Cohort string
-	ID     int
-}
-
-// Voter represents a voter in the election
-type Voter struct {
-	StudentID int
-	Cohort    string
-}
-
-// Vote represents a single vote
-type Vote struct {
-	Hash      string
-	Candidate int
-}
