--- election/hash.go
@@ -1,7 +1,0 @@
-package election
-
-// return some hash of the vote, must be repeatable
-func (v *Vote) hash() {
-	// TODO: implement hash
-
-}
