+++ models/candidate.go
@@ -1,0 +1,8 @@
+package models
+
+// Candidate represents one candidate to vote on
+type Candidate struct {
+	Name   string
+	Cohort string
+	ID     int
+}
