+++ models/voter.go
@@ -1,0 +1,8 @@
+package models
+
+// Voter represents a voter in the election
+type Voter struct {
+	StudentID int
+	Cohort    string
+	Name      string
+}
